+++ plugins/filament_view/core/systems/derived/skybox_system.h
@@ -69,7 +69,7 @@ class SkyboxSystem : public System {
     SkyboxSystem& operator=(const SkyboxSystem&) = delete;
 
     void onSystemInit() override;
-    void update(float deltaTime) override;
+    void update(double deltaTime) override;
     void onDestroy() override;
     void debugPrint() override;
